:label: update balance type to u64

+++ cairo1-contracts/cairo1.cairo
@@ -1,26 +1,26 @@
 #[contract]
 mod Contract {
     struct Storage {
-        balance: u128,
+        balance: u64,
     }
 
     #[event]
-    fn BalanceIncreased(amount: u128) {}
+    fn BalanceIncreased(amount: u64) {}
 
     #[constructor]
-    fn constructor(initial_balance: u128) {
+    fn constructor(initial_balance: u64) {
         balance::write(initial_balance);
     }
 
     #[external]
-    fn increase_balance(amount: u128) {
+    fn increase_balance(amount: u64) {
         let new_balance = balance::read() + amount;
         balance::write(new_balance);
         BalanceIncreased(new_balance);
     }
 
     #[view]
-    fn get_balance() -> u128 {
+    fn get_balance() -> u64 {
         balance::read()
     }
 }
